Fixing compiliation errors on MacOS.

+++ minethd.cpp
@@ -21,6 +21,7 @@
   *
   */
 
+#include <cstdlib>
 #include <assert.h>
 #include <cmath>
 #include <chrono>
@@ -47,7 +48,7 @@ void thd_setaffinity(std::thread::native_handle_type h, uint64_t cpu_id)
 {
 #if defined(__APPLE__)
 	thread_port_t mach_thread;
-	thread_affinity_policy_data_t policy = { cpu_id };
+	thread_affinity_policy_data_t policy = { (int)cpu_id };
 	mach_thread = pthread_mach_thread_np(h);
 	thread_policy_set(mach_thread, THREAD_AFFINITY_POLICY, (thread_policy_t)&policy, 1);
 #else
